refactor: thin public root modules
